Update generate.sh and event_handlers.sh in finit-d ebuild

+++ sys-apps/finit-d/files/dev/event_handlers.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-if [ -n "$FN" = "zap" ] ; then
+if [ -n "$FN" = "zap" ] && [ -e "$pidfile" ] ; then
 	kill -SIGKILL $(cat "$pidfile")
 elif [ -n "$FN" = "declare" ] ; then
 	echo "declare() is not supported"
